+++ Web/ForgeTrust.RazorWire/assets/src/page-navigation.ts
@@ -7,6 +7,7 @@ interface Window {
         formFailureManager?: unknown;
         pageNavigationManager?: unknown;
         sectionCopyManager?: unknown;
+        formInteractionsManager?: unknown;
     };
 }